+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 This project gets auto released on every change to the [Files.com API](https://developers.files.com).
 Auto generated releases contain additions and fixes to models and method arguments, theses will not be documented here.
 
+## [2.0.38-beta] - 2021/12/16
+### Add
+- `file_migration.Client.Wait` poll async `file.FileAction` for results. 
+- `file_migration.Client.LogIterator` given a completed file migration this returns a detailed log of each file event.
+
 ## [2.0.36-beta] - 2021/12/03
 ### Add
 - Alias `folder.Client.ListFor` `file.Client.ListFor`